requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# MuJoCo physics simulation
mujoco>=2.3.0
mujoco-python-viewer>=1.0.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment tracking and logging
wandb>=0.15.0
tensorboard>=2.10.0

# Development and utilities
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0

# Optional RL components
stable-baselines3>=2.0.0
gymnasium>=0.26.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0